requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
anyio==3.6.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
asyncer==0.0.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
charset-normalizer==2.1.0
click==8.0.3
coloredlogs==15.0.1
commonmark==0.9.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==0.10.0
fastapi==0.72.0
filelock==3.8.0
filetype==1.0.9
flatbuffers==2.0
frozenlist==1.3.1
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.27
h11==0.13.0
humanfriendly==10.0
idna==3.3
imageio==2.21.1
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipywidgets==8.0.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.10.3
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-widgets==3.0.0
llvmlite==0.39.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mpmath==1.2.1
multidict==6.0.2
nest-asyncio==1.5.5
networkx==2.8.5
numba==0.56.0
numpy==1.22.3
onnxruntime==1.12.0
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pydantic==1.9.2
pydeck==0.7.1
Pygments==2.13.0
PyMatting==1.1.7
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pyzmq==23.2.1
rembg==2.0.23
requests==2.28.1
rich==12.5.1
scikit-image==0.19.1
scipy==1.8.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.4.0
starlette==0.17.1
streamlit==1.12.0
sympy==1.10.1
tifffile==2022.8.12
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.62.3
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
uvicorn==0.17.0
validators==0.20.0
watchdog==2.1.7
wcwidth==0.2.5
widgetsnbextension==4.0.0
yarl==1.8.1
zipp==3.8.1
